Move MetricsPrinter to core.
